[feat] TopLogoView 추가 (기존 MenuCategoryView에 구현되어있던 상단 로고를 TopLogoView로 분리하여 구현)
