+++ CoCafe/Views/MenuCategoryView.swift
@@ -9,18 +9,6 @@ import UIKit
 
 class MenuCategoryView: UIView {
     
-    private let logoImageView: UIImageView = {
-        let imageView = UIImageView(image: UIImage(named: "Logo"))
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-    
-    private let conanFaceImageView: UIImageView = {
-        let imageView = UIImageView(image: UIImage(named: "ConanLogo"))
-        imageView.contentMode = .scaleAspectFit
-        return imageView
-    }()
-    
     private let menuSegmentedControl: UISegmentedControl = {
         let items = ["Coffee", "Non-Coffee", "Dessert"]
         let control = UISegmentedControl(items: items)
@@ -80,28 +68,11 @@ class MenuCategoryView: UIView {
         }, for: .valueChanged)
         
         // 오토레이아웃 설정
-        logoImageView.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(logoImageView)
-        
-        conanFaceImageView.translatesAutoresizingMaskIntoConstraints = false
-        addSubview(conanFaceImageView)
-        
         menuSegmentedControl.translatesAutoresizingMaskIntoConstraints = false
         addSubview(menuSegmentedControl)
         
         NSLayoutConstraint.activate([
-            logoImageView.centerXAnchor.constraint(equalTo:self.centerXAnchor),
-            logoImageView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 10),
-            //logoImageView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            //logoImageView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            logoImageView.heightAnchor.constraint(equalToConstant: 40),
-            logoImageView.widthAnchor.constraint(equalToConstant: 150),
-            
-            conanFaceImageView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 10),
-            conanFaceImageView.trailingAnchor.constraint(equalTo: logoImageView.leadingAnchor, constant: 5),
-            conanFaceImageView.heightAnchor.constraint(equalToConstant: 40),
-            
-            menuSegmentedControl.topAnchor.constraint(equalTo: logoImageView.bottomAnchor, constant: 10), // 상단에서 10pt 아래
+            menuSegmentedControl.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: 10), // 상단에서 10pt 아래
             menuSegmentedControl.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 10),
             menuSegmentedControl.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -10),
             menuSegmentedControl.heightAnchor.constraint(equalToConstant: 45)
